+++ assets/js/init.js
@@ -2,7 +2,7 @@ import 'matchmedia-polyfill';
 import 'matchmedia-polyfill/matchMedia.addListener';
 import 'srcdoc-polyfill';
 
-import hljs from 'vendor/highlight';
+import hljs from '@/vendor/highlight';
 
 import * as base from './base';