+++ scripts/compile-sprites.js
@@ -3,8 +3,8 @@
 'use strict';
 
 const fs = require('fs');
-
 const path = require('path');
+
 const chalk = require('chalk');
 
 const generateSprites = require('../lib/utils/sprites');