+++ test/clientjs/search.test.js
@@ -1,5 +1,6 @@
 import lunr from 'lunr';
-import * as search from 'search';
+
+import * as search from '@/search';
 
 import { createXHRmock } from './utils';