+++ test/js/testGetMarkdown.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const assert = require('assert');
+
 const MarkdownIt = require('markdown-it');
 
 const markdown = require('../../lib/utils/getMarkdown');