+++ test/js/testGroupName.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const assert = require('assert');
+
 const sinon = require('sinon');
 
 const groupName = require('../../lib/utils/groupName');