+++ test/js/testRender.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const fs = require('fs');
-
 const assert = require('assert');
 const path = require('path');
+
 const del = require('del');
 const Promise = require('bluebird');