+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## UNRELEASED
 
 - 💥 BREAKING: Drop support for Node < 10
-- 💥 BREAKING: Switch to Dart Sass throughout and remove `sass.implementation`
-  option
+- 💥 BREAKING: Switch to Dart Sass throughout, and use as default
+  `sass.implementation` option
 - 💥 BREAKING: Convert Herman-specific options to camelCase (`fontPath`,
   `nunjucks.templatePath`, `sass.jsonFile`, `sass.includePaths`,
   `sass.outputStyle`)
@@ -14,8 +14,13 @@
   in `@example scss` annotations with
   [`sass.use`](https://www.oddbird.net/herman/docs/configuration#sass-use)
   option
+- 🚀 NEW: Default Sass importer (used by `@example scss`) now supports `~`
+  imports of external modules in [Yarn 2](https://yarnpkg.com/) PnP environments
+- 🚀 NEW: Allow using custom Sass importer with `@example scss` via
+  new `sass.importer` option
 - 📝 DOCS: Fix broken links in Changelog
 - 🏠 INTERNAL: Upgrade dependencies:
+  - Switch to [Yarn 2](https://yarnpkg.com/)
   - Remove `gulp-imagemin` and `svgo` --
     [343](https://github.com/oddbird/sassdoc-theme-herman/issues/343)
   - Replace `through2` with Node's native `readable-stream`